small changes to design of program
